+++ desktop/cmp/input/Select.js
@@ -81,6 +81,11 @@ export class Select extends HoistInput {
         openMenuOnFocus: PT.bool,
 
         /**
+         * True to show a "clear" button at the right of the control.  Defaults to false.
+         */
+        enableClear: PT.bool,
+
+        /**
          * Preset list of options for selection. Objects must contain a `value` property; a `label`
          * property will be used for the default display of each option. Other types will be taken
          * as their value directly and displayed via toString().  See also `queryFn` to  supply
@@ -153,6 +158,7 @@ export class Select extends HoistInput {
                 formatOptionLabel: this.formatOptionLabel,
                 isDisabled: props.disabled,
                 isMulti: props.enableMulti,
+                isClearable: props.enableClear,
                 menuPlacement: withDefault(props.menuPlacement, 'auto'),
                 noOptionsMessage: this.noOptionsMessageFn,
                 openMenuOnFocus: props.openMenuOnFocus,